fix a "warning: format ‘%ld’ expects type ‘long int’, but argument 7 has type ‘pool_pointer’"
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2832 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/tex/dumpdata.h
@@ -51,7 +51,7 @@ extern void do_zundump (char *, int, int, FILE *);
 	if ((&(base))[i] < (low) || (&(base))[i] > (high)) {		\
 	  FATAL5 ("Item %u (=%ld) of .fmt array at %lx <%ld or >%ld",	\
                 i, (unsigned long) (&(base))[i], (unsigned long) &(base), \
-                (unsigned long) low, (integer) high);			\
+                (unsigned long) low, (unsigned long) high);			\
       }                                                                 \
     }									\
   } while (0)
